+++ CLI/files/PostTagFileReader.hpp
@@ -19,7 +19,7 @@ class PostTagFileReader : public std::ifstream {
   std::vector<bool> read_prefixed_bits();
 
   // sequence of <= 64 bits written as a 64-bit decimal integer
-  std::vector<bool> read_bits_u64(uint64_t bit_count);
+  std::vector<bool> read_bits_u64(uint8_t bit_count);
 };
 
 #endif  // CLI_FILES_POSTTAGFILEREADER_HPP_
